Reduce call stack pressure in rewriteUrls
In certain processing environments the rapid iteration over anchor
elements in rewriteUrls can overflow the call stack, resulting in a
RangeError.  This patch wraps the finsihed() calls in rewriteUrls in
setImmedate callbacks in order to defer them until later in the event
loop, thereby reducing the pressure on the call stack and allowing
mwoffliner to run to completion.

Fixes #54.