Shrink docker image by 200MB+
The biggest change is that all of the downloaded source tarballs
are now deleted after the code is built and installed. The flattening
of build steps will keep the downloaded layers smaller too.

We also need to install make and g++ since the zimwriterfs image now
uninstalls them.

git is apparently required by npm when installing mwoffliner.