Respect absolute paths for output and cache dirs
We use 'path.join' to expand the path for the output and cache
directories even when an absolute path was given. This results
in the absolute path being joined with the current directory. As
a consequence, a chain of folders as mentioned in the absolute
path are created relative to the current directory. This
makes no sense as absolute paths should be treated as it is.

Correct this by correctly identifying absolute paths and expanding
only relative paths with respect to the current directory.