Better WebP polyfill
The Webp Hero polyfill has a major drawback: it can only convert a WebP at once.
It seems to be a limitation of the emscripten-based libwebp but Webp Hero doesn't do
much about it except throwing an exception.
Given the `polyfillDocument()` method iterates over all `<img />` of the document
to request convertion, it can be very frequent that images are not decoded and thus
just skipped.

The actual behavior is unpredictable as it depends a lot on how quickly the DOM is parsed
by the browser. I believe using lazy loading really helped in the case of mwoffliner
as images are frequently scattered around the page and thus not loaded alltogether.

This replaces the use of the WebPMachine from Webp Hero by a custom one that decodes
one by one (using Webp Hero's code) using a pending list.
It's quite naive but seems to do the job.

This Handler's code currently reside in a gist and is thus downloaded from there.
I've noticed there is no other such dependency in mwoffliner, so if you think it should
be published as a regular npm lib, we can go down this extra road. Let me know