Avoid reducing redirects directly to an Object to avoid memory issue
Reducing the list of redirects directly to an Object caused abnormal
memory consumption. While the exact root cause is unknown, reducing
first to a Map and then transforming into a Object removes the memory
issue. Since Object keys are redirect titles which might be quite big
and there might be many in some occasion, we might have memory
fragmentation issues or something else.