Handle bad errors while downloading an article and replace with placeholder
This commit allows to handle errors happening while downloading an
article or its dependencies in a generic manner, and replace it with a
placeholder HTML explaining what happened.

Placeholder page is based on new design for error pages in kiwix-serve.

Messages on placeholder page are ready to be internationalized.

Errors linked to deleted articles are handled as well with this new
mechanims (hence removing dead links which happened before this).

All these articles are counted as failed articles, and we keep of list
of their IDs for display at the end of the crawl. A maximum threshold is
implemented so that we fail the crawl should we have too many bad
articles.