Keep `typeof` HTML attribute and stop adding custom thumbiner HTML around figures
The `typeof` HTML attribute must be kept because it is used for some
CSS rules.

Previous scraper logic was adding custom HTML around figures and spans.
This does not seem to be necessary anymore with `ActionParse` API, and
even leads to visual glitches.

Both changes are done only for `ActionParse` API since it seems they
were needed with other renderers